+++ AutoTouchPlus.lua
@@ -5,8 +5,9 @@ local _string_converters = {}
 local _metamethods = {
 '__add', '__sub', '__mul', '__div', '__idiv', '__pow',
 '__mod', '__concat', '__index', '__newindex', '__call',
-'__pairs', '__ipairs', '__tostring', '__len',
-'__mode', '__metatable', '__gc'
+'__pairs', '__ipairs', '__tostring', '__len', '__unm',
+'__mode', '__metatable', '__gc', '__eq', '__lt', '__gt',
+'__band', '__bor', '__bxor', '__bnot', '__shl', '__shr'
 }
 
 local classes = {}
@@ -174,11 +175,11 @@ end
 return false
 end
 
-local _print = log or print
+____print = log or print
 function print(...)
 local strings = {}
 for i, v in pairs({...}) do strings[#strings + 1] = str(v) end
-_print(table.concat(strings, '\t'))
+return ____print(table.concat(strings, '\t'))
 end
 
 function pprint(tbl)
@@ -206,32 +207,31 @@ end
 
 
 function _string_converters.class_repr(cls)
-local obj_name, value = 'class', tostring(cls)
+local obj_name, value
 if getmetatable(cls) and getmetatable(cls).__name then
 obj_name = 'instance'
 value = cls.__base_repr or tostring(getmetatable(cls))
+else
+obj_name, value = 'class', tostring(cls)
 end
 return '<'..string.gsub(value, 'table:', cls.__name..' '..obj_name..' at')..'>'
 end
 
 function _string_converters.class_str(cls)
-if list{'dict', 'list', 'set'}:contains(getmetatable(cls).__name) then
+if set{'dict', 'list', 'set'}:contains(getmetatable(cls).__name) then
 return _string_converters.table2string(cls)
-else
-return _string_converters.class_repr(cls)
 end
+return _string_converters.class_repr(cls)
 end
 
 
 function _string_converters.table2string(input)
 local m = getmetatable(input)
 local function idxstr(idx, val, custom_type)
-if custom_type then return str(val)
-elseif is.str(val) then val = '"'..val..'"'
-else val = str(val) end
-if is.str(idx) then idx = '"'..idx..'"'
-else idx = str(idx) end
-return string.format('%s: %s', idx, val)
+if custom_type then return str(val) end
+if is.str(val) then  val = '"'..val..'"'  end
+if is.str(idx) then idx = '"'..idx..'"' end
+return string.format('%s: %s', str(idx), str(val))
 end
 
 local custom = m and m.__name and list{'list', 'set'}:contains(m.__name)
@@ -273,7 +273,6 @@ local typeSortOrder = {
 ['userdata'] = 7;
 ['nil']      = 8;
 }
-
 
 local function isPrimitiveArray(array)
 local max,n = 0,0
@@ -2238,6 +2237,7 @@ function Triangle:__tostring()
 return string.format('<Triangle(n=%d, color=%d)>', len(self.pixels), self.color)
 end
 
+
 requests = {}
 
 function requests.delete(url, args) return requests.request("DELETE", url, args) end
@@ -2258,21 +2258,28 @@ log.debug('Sending %s request: %s with data: %s', _req.method, _req.url or _req[
 return request:send(request:build())
 end
 
-local function parse_data(lines, request, response)
-local err_msg = 'error in '..request.method..' request: '
-assert(isnotin('failed', lines[6]), err_msg..'Url does not exist')
 
-for i, ln in pairs(lines) do
-local code, reason = ln:match('HTTP request sent, awaiting response[^%d]*(%d+) (.*)')
-local content_length, mime_type = ln:match('Length: (%d+) %[(.*)%]')
-if code then response.status_code = code end
-if reason then response.reason = reason end
-if content_length then response.content_length = content_length end
-if mime_type then response.mime_type = mime_type end
+local function parse_data(lines, response)
+local before_empty_line = true
+
+for i, ln in pairs(lines(2, nil)) do
+ln = ln:replace('\13', '')
+if before_empty_line then
+if ln == '' then
+before_empty_line = false
+else
+local kv = ln:split(':')
+local k, v = kv[1], (':'):join(list(kv)(2, nil)):strip(' ')
+if v == 'true' then v = true elseif v == 'false' then v = false end
+response.headers[k] = tonumber(v) or v
 end
-
-response.status_code = num(response.status_code)
-response.ok = response.status_code < 400
+else
+response.text = response.text..ln..'\n'
+end
+end
+local info = lines[1]:replace('\13', ''):split(' ')
+response.http_version, response.status_code, response.reason = info[1], info[2], (' '):join(info(3, nil))
+response.ok, response.status_code = response.reason == 'OK', num(response.status_code)
 end
 
 local function urlencode(params)
@@ -2285,6 +2292,7 @@ if tostring(value) then s = s..tostring(key)..'='..tostring(value) end
 end
 return s
 end
+
 
 Request = class('Request')
 function Request:__init(request)
@@ -2304,58 +2312,40 @@ end
 end
 
 function Request:build()
-local cmd = list{'wget', '--method', self.method:upper()}
+local cmd = list{'curl', '-si', '-X', self.method:upper()}
 if is(self.params) then
 self.url = self.url .. '?' .. urlencode(self.params)
 end
-cmd:extend(self:_add_auth() or {})
-cmd:extend(self:_add_data() or {})
-cmd:extend(self:_add_headers() or {})
-cmd:extend(self:_add_proxies() or {})
-cmd:extend(self:_add_ssl() or {})
-cmd:extend(self:_add_user_agent() or {})
-cmd:extend{"'"..self.url.."'"}
-cmd:extend{'--output-file', '-'}
-cmd:extend{'--output-document', self._response_fn}
+for k in iter{'auth', 'data', 'headers', 'proxies', 'ssl', 'user_agent'} do
+cmd:extend(getattr(self, '_add_'..k)(self) or {})
+end
+cmd:append("'"..self.url.."'")
 return cmd
 end
 
 function Request:send(cmd)
 local response = Response(self)
-
 try(function()
 local lines = exe(cmd, true, true)
-
-local response_f = assert(io.open(self._response_fn))
-response.text = response_f:read('*a')
-response_f:close()
-
-try(function() parse_data(lines, self, response) end)
-
+parse_data(lines, response)
 end,
-
 except(function(err)
-log.error('Failed to fetch url: ' .. str(err))
-end),
-
-function()
-exe{'rm', self._response_fn}
-end
-)
-
+log.error('Failed to request url: %s - %s ', self.url, str(err))
+end))
 return response
 end
 
 function Request:verify()
-assert(requal(self.data, json.decode(json.encode(self.data))),'Incorrect json formatting')
-assert(self.url:startswith('http'), 'Only http(s) urls are supported')
+local prefix = 'Invalid request: '
+assert(requal(self.data, json.decode(json.encode(self.data))), prefix..'Incorrect json formatting')
+assert(self.url:startswith('http'), prefix..'Only http(s) urls are supported')
 end
 
 function Request:_add_auth()
 if is(self.auth) then
 local usr = self.auth.user or self.auth[1]
 local pwd = self.auth.password or self.auth[2]
-return {'--http-user', usr, '--http-password', pwd}
+return {'--basic', '--user', "''"..usr..':'..pwd.."''"}
 end
 end
 
@@ -2364,7 +2354,7 @@ if is(self.data) then
 if Not.string(self.data) then
 self.data = urlencode(self.data)
 end
-return {'--body-data', "'"..self.data.."'"}
+return {'--data', "'"..self.data.."'"}
 end
 end
 
@@ -2372,30 +2362,31 @@ function Request:_add_headers()
 local cmd = list()
 if is(self.headers) then
 for k, v in pairs(self.headers) do
-cmd:append("--header='"..k..': '..str(v).."'")
+cmd:extend{"--header", "'"..k..': '..str(v).."'"}
 end
 end
 return cmd
 end
 
 function Request:_add_proxies()
-if is(self.proxies) then
-local usr, pwd
-for k, v in pairs(self.proxies) do
-if isin('@', v) then usr, pwd = unpack(v:split('//')[2]:split('@')[1]:split(':')) end
-end
-end
+-- TODO: request with proxy and test
+-- if is(self.proxies) then
+--   local usr, pwd
+--   for k, v in pairs(self.proxies) do
+--     if isin('@', v) then usr, pwd = unpack(v:split('//')[2]:split('@')[1]:split(':')) end
+--   end
+-- end
 end
 
 function Request:_add_ssl()
 if not self.verify_ssl or (self.url:startswith('https') and self.verify_ssl) then
-return {'--no-check-certificate'}
+return {'--insecure'}
 end
 end
 
 function Request:_add_user_agent()
 if is(self.user_agent) then
-return {'-U', self.user_agent}
+return {'--user-agent', "'"..self.user_agent.."'"}
 end
 end
 
@@ -2403,15 +2394,13 @@ Response = class('Response')
 function Response:__init(request)
 assert(request, 'Cannot create response with no request')
 self.request = request or {}
-self.method = self.request.method
-self.url = self.request.url
-self.status_code = -1
-self.reason = ''
-self.text = ''
-self.encoding = 'utf-8'
-self.mime_type = 'text/html'
 self.headers = dict()
 self.ok = false
+self.method = self.request.method
+self.reason = ''
+self.status_code = -1
+self.text = ''
+self.url = self.request.url
 end
 
 function Response:__tostring()
@@ -2429,8 +2418,19 @@ return json.decode(self.text)
 end
 
 function Response:raise_for_status()
-if self.status_code ~= 200 then error('error in '..self.method..' request: '..self.status_code) end
+if self.status_code < 200 or self.status_code >= 400 then
+error('Error in '..self.method..' response: '..self.status_code)
 end
+end
+
+property(Response, 'content_length',
+function(self, k) return self.headers:get('Content-Length', 0) end)
+
+property(Response, 'content_type',
+function(self, k) return self.headers:get('Content-Type', ''):split(';')[1] end)
+
+property(Response, 'encoding',
+function(self, k) return self.headers:get('Content-Type', ''):match('charset=(.*)') end)
 
 screen = {
 before_action_funcs = set(),
@@ -2802,10 +2802,12 @@ log.debug('Copying files from %s to %s', src, dest)
 local cmd = list{'cp'}
 if os.is_dir(src) then cmd:append('-R') end
 if not overwrite then cmd:append('-n') end
+-- luacov: disable
 if add_rootDir ~= false and rootDir then
 src = os.path_join(rootDir(), src)
 dest = os.path_join(rootDir(), dest)
 end
+-- luacov: enable
 cmd:extend{src, dest}
 exe(cmd, true, true)
 end